check_readiness.py
import weaviate
from weaviate.classes.init import Auth
import os

wcd_url = os.environ["WCD_URL"]
wcd_api_key = os.environ["WCD_API_KEY"]

client = weaviate.connect_to_weaviate_cloud(
    cluster_url=wcd_url,
    auth_credentials=Auth.api_key(wcd_api_key),
)

print(client.is_ready())
client.close()
# should print `True`
